Refactored MaterialProps to import from ConstitutiveModels folder
