+++ safeincave/ConstitutiveModels/__init__.py
@@ -1,0 +1,37 @@
+# Copyright (c) 2026, The SafeInCave Developers
+#
+# SPDX-License-Identifier: BSD-3-Clause
+
+"""
+Constitutive models module for SafeInCave.
+
+This module contains all constitutive element classes for material behavior modeling:
+- Elastic: Spring
+- Thermoelastic: Thermoelastic
+- Non-elastic mechanisms: Viscoelastic, LinearDashpot, DislocationCreep, 
+  PressureSolutionCreep, ViscoplasticDesai, MunsonDawsonCreep, ModifiedCamClayViscoplastic
+"""
+
+from .NonElasticElement import NonElasticElement
+from .Thermoelastic import Thermoelastic
+from .Spring import Spring
+from .Viscoelastic import Viscoelastic
+from .LinearDashpot import LinearDashpot
+from .DislocationCreep import DislocationCreep
+from .PressureSolutionCreep import PressureSolutionCreep
+from .ViscoplasticDesai import ViscoplasticDesai
+from .MunsonDawsonCreep import MunsonDawsonCreep
+from .ModifiedCamClayViscoplastic import ModifiedCamClayViscoplastic
+
+__all__ = [
+    "NonElasticElement",
+    "Thermoelastic",
+    "Spring",
+    "Viscoelastic",
+    "LinearDashpot",
+    "DislocationCreep",
+    "PressureSolutionCreep",
+    "ViscoplasticDesai",
+    "MunsonDawsonCreep",
+    "ModifiedCamClayViscoplastic",
+]
